Refactor SetupUsingProxyWithSignIn component to reduce cyclomatic complexity.
